make edit form its own component
